requirements.txt
pandas
requests
beautifulsoup4
pyyaml
cot_reports
plotly
kaleido